Remove unused plugin system with type errors
After fixing some type errors in @hyperjump/json-schema, it exposed a
type issue in the plugin system. We aren't using the plugin system, so I
decided to just remove it for now instead of fix it. Something like it
can be added in again later if it turns out we need it.